implementation of a graphics strategy. based on vector graphics.
using svg. and the embed tag we can merge all of our graphics into a
single file (icons.svg). this will increase the scalability of the
application.

while this is a complete implementation. we are currently still serving
our graphics as individual graphics. will make the switch to a single
graphic file after we've locked down the ui a little more.