starting implementation of copper graph
we mean to kill json.

why? because it's not semantic. which results in developers having to
move their object model into their javascript.

If you think of the user agent as a MVC (which we do) then this is a
serious violation.

this graph should empower us with close to ths same size as json,
better semantics and a much cleaner baseline to build a reactive user
interface.